requirements.txt
numpy
pillow
selenium